+++ cherrypy/lib/profiler.py
@@ -105,23 +105,10 @@ class Profiler(object):
         :rtype: str
         """
         sio = io.StringIO()
-        if sys.version_info >= (2, 5):
-            s = pstats.Stats(os.path.join(self.path, filename), stream=sio)
-            s.strip_dirs()
-            s.sort_stats(sortby)
-            s.print_stats()
-        else:
-            # pstats.Stats before Python 2.5 didn't take a 'stream' arg,
-            # but just printed to stdout. So re-route stdout.
-            s = pstats.Stats(os.path.join(self.path, filename))
-            s.strip_dirs()
-            s.sort_stats(sortby)
-            oldout = sys.stdout
-            try:
-                sys.stdout = sio
-                s.print_stats()
-            finally:
-                sys.stdout = oldout
+        s = pstats.Stats(os.path.join(self.path, filename), stream=sio)
+        s.strip_dirs()
+        s.sort_stats(sortby)
+        s.print_stats()
         response = sio.getvalue()
         sio.close()
         return response